move ansi support out of env
